Update and rename rectangleorsquare.java to Rectangleorsquare.java
class name in uppercase.
evaluation of rectangle or square in function
